Input with all relevant files
Work is still being done on the input, so ignore the fact it says 'final'. This should extract the data, ensure there are no errors and output everything to a text file. For it to work, please just execute the code in C# (don't try to get the C# to interact with it further than that) and then read everything from the text documents. Heads up, when you execute the code, you may need to pass in variables, like you would if you were running any form of method or function. May need some co-ordination between python and C# so let me know if you have any trouble. Also, make sure your laptop is connected to the internet else it won't run. Will be uploading how to download what you need shortly.